tests passed

+++ libft/ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: gsuhr <gsuhr@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/15 12:50:51 by gsuhr             #+#    #+#             */
-/*   Updated: 2023/12/14 18:36:24 by gsuhr            ###   ########.fr       */
+/*   Updated: 2023/12/20 16:26:09 by gsuhr            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *str)
+#include "libft.h"
+
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
